+++ src/main/kotlin/com/dobby/backend/application/usecase/experiment/UpdateMyExperimentPostRecruitStatusUseCase.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend.application.usecase.member
+package com.dobby.backend.application.usecase.experiment
 
 import com.dobby.backend.application.usecase.UseCase
 import com.dobby.backend.domain.exception.ExperimentPostNotFoundException
