+++ src/test/kotlin/com/dobby/backend/application/usecase/experiment/UpdateMyExperimentPostRecruitStatusUseCaseTest.kt
@@ -1,5 +1,6 @@
-package com.dobby.backend.application.usecase.member
+package com.dobby.backend.application.usecase.experiment
 
+import com.dobby.backend.application.usecase.experiment.UpdateMyExperimentPostRecruitStatusUseCase
 import com.dobby.backend.domain.exception.ExperimentPostNotFoundException
 import com.dobby.backend.domain.exception.ExperimentPostRecruitStatusException
 import com.dobby.backend.domain.gateway.experiment.ExperimentPostGateway
